// src/types/chart/chart.ts
export interface DatasetDesign {
  readonly backgroundColor?: string;
  readonly yAxisId?: string;
  readonly label?: string;
}

export interface AxisDesign {
  readonly title?: TitleDesign;
}

export interface AxesDesign {
  readonly x?: AxisDesign;
  readonly y?: AxisDesign;
  readonly y1?: AxisDesign;
}

export interface TitleDesign {
  readonly text?: string;
  readonly bold?: boolean;
  readonly italic?: boolean;
  readonly align?: Align;
  readonly color?: Color;
}

export type TrendType = "polynomial" | "exponential" | "logarithmic" | "trailingMovingAverage";
export interface TrendConfiguration {
  type?: TrendType;
  order?: number;
  color?: Color;
  display?: boolean;
  window?: number;
}

export type CustomizedDataSet = {
  readonly dataRange: string;
  readonly trend?: TrendConfiguration;
} & DatasetDesign;

export type AxisType = "category" | "linear" | "time";

export interface DataSet {
  readonly labelCell?: Range; // range of the label
  readonly dataRange: Range; // range of the data
  readonly rightYAxis?: boolean; // if the dataset should be on the right Y axis
  readonly backgroundColor?: Color;
  readonly customLabel?: string;
}
export interface ExcelChartDataset {
  readonly label?: { text?: string } | { reference?: string };
  readonly range: string;
  readonly backgroundColor?: Color;
  readonly rightYAxis?: boolean;
}

export type ExcelChartType = "line" | "bar" | "pie" | "combo" | "scatter" | "radar";

export interface ExcelChartDefinition {
  readonly title?: TitleDesign;
  readonly type: ExcelChartType;
  readonly dataSets: ExcelChartDataset[];
  readonly labelRange?: string;
  readonly backgroundColor: XlsxHexColor;
  readonly fontColor: XlsxHexColor;
  readonly legendPosition: LegendPosition;
  readonly stacked?: boolean;
  readonly cumulative?: boolean;
  readonly verticalAxis?: {
    useLeftAxis?: boolean;
    useRightAxis?: boolean;
  };
  readonly axesDesign?: AxesDesign;
}

export interface ChartCreationContext {
  readonly range?: CustomizedDataSet[];
  readonly title?: TitleDesign;
  readonly background?: string;
  readonly auxiliaryRange?: string;
  readonly aggregated?: boolean;
  readonly stacked?: boolean;
  readonly cumulative?: boolean;
  readonly dataSetsHaveTitle?: boolean;
  readonly labelsAsText?: boolean;
  readonly showSubTotals?: boolean;
  readonly showConnectorLines?: boolean;
  readonly firstValueAsSubtotal?: boolean;
  readonly legendPosition?: LegendPosition;
  readonly axesDesign?: AxesDesign;
  readonly fillArea?: boolean;
  readonly showValues?: boolean;
}